+++ src/app/styles/body.scss
@@ -1,0 +1,3 @@
+.no-scroll {
+  overflow: hidden; /* Bloquea el scroll */
+}
